requirements.txt
flask==2.1.1
flask_cors==4.0.0
fastapi==0.100.0
requests==2.31.0
openpyxl==3.1.2
python-telegram-bot==13.13
uvicorn==0.22.0
httpx==0.24.1